refactor: 관리자 전시 페이지 Header 컴포넌트화 및 졸업 전시 페이지 내 적용 (#107)
